refactor: use cli-skeleton for commands
This improves consistency across Dokku-related tools.
