+++ .gitignore
@@ -22,3 +22,5 @@ validation/*
 
 # .env files
 .env*
+
+dokku-event-listener
